split up all the snesing processing to be easier to manage
